Fixes #243. Also fixes #300. This incorporates suggestions by Alex to not make the 'generator menu' disappear but clicking it when an error is present will make it throw out an error in the console.
