+++ bundles/org.eclipse.vorto.codegen.ui/src/org/eclipse/vorto/codegen/ui/handler/CodeGeneratorInvocationHandler.java
@@ -14,14 +14,22 @@
  *******************************************************************************/
 package org.eclipse.vorto.codegen.ui.handler;
 
+import java.util.Collection;
+import java.util.Collections;
+
 import org.eclipse.core.commands.AbstractHandler;
 import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
 import org.eclipse.core.runtime.IConfigurationElement;
+import org.eclipse.emf.common.util.Diagnostic;
+import org.eclipse.emf.ecore.EObject;
+import org.eclipse.emf.ecore.util.Diagnostician;
 import org.eclipse.vorto.codegen.api.IVortoCodeGenerator;
 import org.eclipse.vorto.codegen.api.mapping.InvocationContext;
 import org.eclipse.vorto.codegen.ui.utils.PlatformUtils;
+import org.eclipse.vorto.core.api.model.informationmodel.FunctionblockProperty;
 import org.eclipse.vorto.core.api.model.informationmodel.InformationModel;
+import org.eclipse.vorto.core.api.model.model.Model;
 import org.eclipse.vorto.core.ui.MessageDisplayFactory;
 import org.eclipse.vorto.core.ui.model.IModelElement;
 import org.eclipse.vorto.core.ui.model.IModelProject;
@@ -46,21 +54,29 @@ public class CodeGeneratorInvocationHandler extends AbstractHandler {
 	@Override
 	public Object execute(ExecutionEvent event) throws ExecutionException {
 		final String generatorIdentifier = event.getParameter("org.eclipse.vorto.codegen.generator.commandParameter");
-		evaluate(generatorIdentifier);
-		PlatformUtils.switchPerspective(PlatformUtils.JAVA_PERSPECTIVE);
+		if (evaluate(generatorIdentifier)) {
+			PlatformUtils.switchPerspective(PlatformUtils.JAVA_PERSPECTIVE);
+		}
+
 		return null;
 	}
 
-	private void evaluate(String generatorName) {
+	private boolean evaluate(String generatorName) {
 
 		final IConfigurationElement[] configElements = getUserSelectedGenerators(generatorName);
-		
-		
+
 		IModelElement selectedElement = ModelProjectFactory.getInstance().getModelElementFromSelection();
 		if (selectedElement == null) {
-			MessageDisplayFactory.getMessageDisplay().displayWarning("Model was not properly selected. Please try again.");
+			MessageDisplayFactory.getMessageDisplay()
+					.displayWarning("Model was not properly selected. Please try again.");
+			return false;
 		}
-		
+
+		if (!hasNoErrors(selectedElement)) {
+			MessageDisplayFactory.getMessageDisplay().displayError("Model has errors. Cannot generate.");
+			return false;
+		}
+
 		InformationModel informationModel = (InformationModel) selectedElement.getModel();
 
 		for (IConfigurationElement e : configElements) {
@@ -69,19 +85,56 @@ public class CodeGeneratorInvocationHandler extends AbstractHandler {
 
 				// interested only in code generators
 				if (!(codeGenerator instanceof IVortoCodeGenerator)) {
-					continue; 
+					continue;
 				}
 
 				IVortoCodeGenerator informationModelCodeGenerator = (IVortoCodeGenerator) codeGenerator;
 
 				CodeGeneratorTaskExecutor.execute(informationModel, informationModelCodeGenerator,
-						createInvocationContext(selectedElement.getProject(), informationModelCodeGenerator.getServiceKey()));
+						createInvocationContext(selectedElement.getProject(),
+								informationModelCodeGenerator.getServiceKey()));
 
 			} catch (Exception e1) {
 				MessageDisplayFactory.getMessageDisplay().displayError(e1);
 				throw new RuntimeException("Something went wrong during code generation", e1);
 			}
 		}
+
+		return true;
+	}
+
+	private boolean hasNoErrors(IModelElement selectedElement) {
+		return selectedElement.getDiagnostics().size() <= 0 && noLinkingErrors(selectedElement);
+	}
+
+	// check if model has no syntax errors
+	private boolean noLinkingErrors(IModelElement modelElement) {
+		Model model = modelElement.getModel();
+
+		if (model instanceof InformationModel) {
+			InformationModel infoModel = (InformationModel) model;
+			for (FunctionblockProperty property : infoModel.getProperties()) {
+				// for syntax and parsing errors
+				if (property.getType().eResource().getErrors().size() > 0) {
+					return false;
+				}
+
+				// for linking errors
+				if (getLinkingErrors(property.getType()).size() > 0) {
+					return false;
+				}
+			}
+		}
+		return true;
+	}
+
+	private Collection<Diagnostic> getLinkingErrors(EObject model) {
+		Diagnostic diagnostic = Diagnostician.INSTANCE.validate(model);
+		switch (diagnostic.getSeverity()) {
+		case Diagnostic.ERROR:
+			return diagnostic.getChildren();
+		}
+		return Collections.emptyList();
 	}
 
 	private InvocationContext createInvocationContext(IModelProject project, String targetPlatform) {
